Add working single threaded executor with timers

+++ include/publisher.hpp
@@ -10,10 +10,10 @@ namespace rclcpp {
 template <typename T>
 class Publisher {
     const std::string topic_name;
-    double queue_size;
+    int queue_size;
 
    public:
-    Publisher(std::string topic_name, double queue_size) : topic_name{topic_name}, queue_size{queue_size} {}
+    Publisher(std::string topic_name, int queue_size) : topic_name{topic_name}, queue_size{queue_size} {}
     ~Publisher() {};
     void publish(T msg) {
         RMW::publish_message<T>(topic_name, msg);
